Add version 1.0.0
I wrote all this because I didn't quite like dashboard.el, and I felt
like hacking on it enough to like it would have amounted to roughly
the same amount of effort in the end as writing something from scratch.
Because this is something that I wrote for myself, I cut corners in some
places, something I would not do if I was writing for the general public.
Therefore, I would recommend against using the package yourself;
I'm putting it online only because it makes things easier for me. However,
it might still serve as a starting point for writing something similar yourself,
or for a fork adding more features.

The design was inspired by the Doom dashboard.